Refactor dropdown to factor out positionable logic, in preparation of extending to tooltip
